pyproject.toml
[tool.poetry]
name = "lojack-clients"
version = "1.0.0"
description = "A collection of unofficial client libraries for accessing Spireon's LoJack APIs"
readme = "README.md"
packages = [
    { include = "lojack_clients", from = "src" },
]
include = ["src/lojack_clients/*/py.typed"]
keywords = ["lojack", "spireon", "sdk"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries"
]
authors = ["Shawn Corrigan <sha.wnja.mesc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
httpx = ">=0.15.4,<0.23.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
uuid = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py37', 'py38', 'py39']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
